Update patch set 1
Patch Set 1:

hlte is work in progress; but there are already deployed aboots requiring v2 in a couple of the variants.

and yes, it would. but the reason we don't have a common copy is because different devices require different versions. I know of at least 2 upcoming devices that require v2, not including the hlte. Unless we name this one differently to avoid conflicts, we can't merge this patch.

The ideal scenario would be to make a version that defaults to v1 and accepts a flag to use v2, and switch existing hardware to that.

Patch-set: 1