Update patch set 4
Patch Set 4:

We should really just make a kernel patch which standardizes the interface at the kernel level. We get a lot of randomness from vendor code and I think we could stand to reduce the chaos in some of these cases. I ended up creating an interface like this for display stuff after getting sick of seeing the insane stuff that OEMs do.

Patch-set: 4