Update patch set 2
Patch Set 2:

The perfd client is used in more places than the PowerHAL, for example it's loaded into mediaserver which is 32-bit only on 64-bit devices (hence loading the 32-bit library). The PowerHAL on the other hand is loaded by the 64-bit system server via JNI, which in turn loads the 64-bit perfd client.

Patch-set: 2