Update patch set 2
Patch Set 2:

Also in CM, neither the perfd client nor the PowerHAL is loaded into applications. If an application requests a boost, it goes thru the API in CMSDK, which invokes the binder server in the framework, which calls into JNI, which calls to the PowerHAL, which sends a message to perfd. Whew.

This is not the case for the Qualcomm Boost Framework though, which we do not use.

Patch-set: 2