Update patch set 4
Patch Set 4:

I meant how updated sony adreno blobs are, and whether or not we can commonize a2xx blobs. However, weird that huashan is 8960 (DC) and a3xx, I'll put that up to some drunk Sony engineer.
Either way, I think this should be renamed to 8064, as the majority (all but huashan afaik) of 8960 devices that have a3xx are 8064

Patch-set: 4