Update patch set 1
Patch Set 1:

I'd need to look into it again... I'm using a nonstandard libqti-perfd-client, which should probably be swapped for the xiaomi one or something more CAF standard.

Also, are we intending to follow upstream CAF for the power HALs? If so, we might want to consider doing a better job of organizing them in our qcom/common, or somehow making our qcom/common more upstream mergeable.

Patch-set: 1