Update patch set 1
Patch Set 1:

If we really believe that these are qcom supplied blobs then I am good with merging the HAL commits and forcing the blob update on the devices.

I have been running with the two topics

caf-8996-03400
caf-8996-02500

for the last 20 hours and haven't encountered any problems.  Shall we move forward with it now that we all have a little time before next week's builds start?

Patch-set: 1